modified for negative weight

+++ includes/rb_tree.h
@@ -6,7 +6,7 @@
 /*   By: aulopez <aulopez@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/31 14:56:35 by aulopez           #+#    #+#             */
-/*   Updated: 2019/07/17 13:36:15 by aulopez          ###   ########.fr       */
+/*   Updated: 2019/07/27 20:04:47 by aulopez          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -50,7 +50,7 @@ typedef struct			s_rb_node
 	size_t				visited;
 	t_list				*link;
 	char				*name;
-	uint8_t				flag;
+	size_t				flag;
 }						t_rb_node;
 
 void					rb_balance(t_rb_node **node, int am_i_left);
